+++ Client/Math.cpp
@@ -17,7 +17,7 @@ void Math::Normalize(float* fX, float* fY)
 
 float Math::Rotate(GameObject* gGameObject, float vLocalPositionX, float vLocalPositionY)
 {
-	return -atan2(vLocalPositionX - gGameObject->m_fX, vLocalPositionY - gGameObject->m_fY) * 180 / 3.14159;
+	return static_cast<float>(- atan2(vLocalPositionX - gGameObject->m_fX, vLocalPositionY - gGameObject->m_fY) * 180 / 3.14159);
 }
 
 bool Math::IsInsideInterval(float v, float vMin, float vMax)
