+++ Serveur2/GameManagerS.cpp
@@ -80,7 +80,7 @@ void GameManager::SendJSON(bool GameRunnig, bool ValidMove) {
         m_jServ["Grid"] = m_Grid;
         std::string jtext = m_jServ.dump() + "\n";
         // send json to server
-        int bytesSent = send(pPlayer->m_sSock, jtext.c_str(), strlen(jtext.c_str()), 0);
+        int bytesSent = send(pPlayer->m_sSock, jtext.c_str(), static_cast<int>(strlen(jtext.c_str())), 0);
         if (bytesSent == SOCKET_ERROR)
         {
             if (WSAGetLastError() != WSAEWOULDBLOCK)
